requirements.txt
line-bot-sdk
flask
configparser
azure-ai-translation-text
azure-cognitiveservices-speech
librosa
